Remove the useless "Expand Graph" button.
Instructions for Claude:
The "Expand Graph" button doesn't seem to actually do anything. What was
it supposed to do?

A node cannot be selected, because clicking on that node automatically
causes that node to be expanded. So that button appears to be useless.
Is that correct?

Please remove that button. Any unused code called by that button should
also be removed.