Allow draing of large graphs to be halted.
Instructions for Claude:
If there are more than several hundred Atoms to be drawn, the user should
be warned that the graph may be large and may take a long time to draw.
If fetching and drawing is taking more than a few seconds, there should
be a button displayed that allows the user to stop the processing. After
stopping the processing, the viewer should be left in a usable state,
and continue to function normally, if the user changes the layout.