Fix the "Stop Processing" Button
Instructions for Claude:
The "Stop Processing" button that you just added does not work quite
right in the graph view. I tried to download a large graph, and I
clecked on it. Instead of stopping the processing, it continued to
perform queries for the incoming set of each ListLink. Not only that,
but it also tried to redraw the graph after receiving each new EdgeLink
as a result of those queries. The "Stop Processing" button should halt
the JSON traffic as well, since there might be thousands or millions
of ListLinks.