Implement max-cache size
Instructions for Claude:
Please limit the size of the atom-cache to be no larger than 250 Atoms.
The goal is to avoid excessively long waits during redraws. If the user
has triggered a fetch of an Incoming set that would cause the size of
the cache to grow larger than this, add as many links as you can, as
long as the max size of the cache is not exceeded. Do not add them
if the max size is exceeded.

If the result of the operation is that not all of the Links were added
to the cache, then a warning should be printed at the top of the page.
That warning should stay in place as long as the cache is near full.
Because the size of the cache can decrease when the user removes arrows,
the warning message should go away when the cache size is less than 96%
of this max limit. It should come back, if the limit is approached or
exceeded again.

Please make this maximum-size parameter configurable, with a box in the
tree-view.html page. Also, please show the total number of Atoms in
currently in the atomspace-cache. This total should be updated whenever
atoms are added to or removed from the cache.