Fix link deletion in graph view.
Instructions for Claude:
In the graph-layout view, clicking on an arrow causes Atoms to be
removed. Currently, it seems that both the atoms at the head and the
fail are removed. Please change this so that the only the tail Atom is
removed. The head (the atom being pointed at) should stay.