(Again) Fix exceessive redraws when cache-max is hit.
Instructions for Claude:
Please go to the cogserver directory and examine the javascript
files in examples/visualizer.

In the graph view, when fetching the incoming set for a node that has
a very large incoming set, the cache max is quickly reached. When this
happens, the atomspace cache contains many ListLinks that are not
contained in any EdgeLinks. This will be drawn as just single
unconnected vertexes in the graph view. This is correct. The graph
redraw is fast, because when the cache-max is hit, the reports are
batched, and the graph is redrawn nly once, at the end.  However,
it the cache size is then manually increased to be larger, then
there will be an attempted redraw for each now EdgeLink that arrives.
There might be many hundreds of redraws, with several attempted every
second. These redraws should be held off until the end, when the
atomsspace cache has finished fetching everything it needs.