(Again) Increasing cache does not resume operations
Instructions for Claude:
The most recent change made things worse.  Now, there are multiple
redraws, even when loading atoms for the first time. This should not
happen, and it did not happen before.  In addition, when increasing
the cache size, while the atomspace-cache is still updating, multiple
redraws happen in rapid succsession. This should not happen, and it
did not happen in the previous version. The third problem is that when
the AtomSpace cache settles down, then increasing the cache size does
not trigger a redraw. It should.